requirements.txt
gunicorn==21.2.0
psycopg2-binary==2.9.9
Bootstrap-Flask==2.3.3
Flask==3.0.2
Pillow==10.2.0
numpy==1.26.4